+++ iManagerGen/IManagerGenNHibernate/CEN/IManager/PedidoCEN.cs
@@ -95,7 +95,7 @@ public System.Collections.Generic.IList<PedidoEN> DameTodos (int first, int size
         list = _IPedidoCAD.DameTodos (first, size);
         return list;
 }
-public int ContarPedidosPendientes ()
+public long ContarPedidosPendientes ()
 {
         return _IPedidoCAD.ContarPedidosPendientes ();
 }
