+++ iManagerGen/IManagerGenNHibernate/CEN/IManager/UsuarioCEN.cs
@@ -73,7 +73,7 @@ public System.Collections.Generic.IList<UsuarioEN> DameTodos (int first, int siz
         list = _IUsuarioCAD.DameTodos (first, size);
         return list;
 }
-public int ContarUsuarios ()
+public long ContarUsuarios ()
 {
         return _IUsuarioCAD.ContarUsuarios ();
 }
